Assigned different learning rates to different layers.
The conv layers have smaller rates, while the fc layers and segmenter layer have larger rates.
We still have lr_scheduling, though it is modified to take into account various learning rates for different layers.

As a result, after the first epoch, we get an m-IoU of 27% which is about 5-7 percent higher than the old case where all the layers had similar learning rates.

Note that we are working on train_v2 and train.py is still on the old form.