Updated train_v2.py to work for multiple cases:
[1] suprevised pre-trained case (pre_trained == 'gt')
[2] self-supervised pre-trained case (pre_trained == 'self')
[3] Np pre-trained model (pre_trained == 'no')

Right now, just tested case [1] and after one epoch, I get m-IoU=32% which is really good.
To be tested to see if the code works: [3] and then [2].

Moreover, pascal_voc_loader was updated to take into account the normalization when the pre-trained model is not built on image-net data. In that case I have simply normalized data by mean(0.5,0.5,0.5) & std(0.5,0.5,0.5).