Added dcgan_alexnet that feasts on Imagenet dataset.
The discriminator in this GAN (netD) is of type AlexNetGAN, defined in networks.py. AlexNetGAN is a fully convolutional AlexNet where Maxpooling layers are replaced by strides of two in the convolution layers. It also features Batchnorms and LeakyRelues.

Once the GAN is trained for 1-2 epochs over Imagenet, we can check how good the trained alexnet works as pre-training for our semantic labeling purpose.
Note that we need to modify the definitoin of our previously defined AlexFCNv2 to make it compatible with the structure of AlexNetGAN.